Meme-Generator/requirements.txt
altair==1.2.1
asn1crypto==0.22.0
atari-py==0.1.7
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.6
av==0.3.3
awscli==1.16.17
backcall==0.1.0
backports.functools-lru-cache==1.4
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bleach==1.5.0
blinker==1.4
bokeh==0.12.13
boto==2.48.0
boto3==1.9.7
botocore==1.12.7
Box2D==2.3.2
Box2D-kengz==2.3.3
bresenham==0.2
bz2file==0.98
cassandra-driver==3.11.0
certifi==2019.11.28
cffi==1.11.2
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
colorama==0.3.9
colour==0.1.5
conda==4.6.14
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.7
cytoolz==0.9.0.1
dask==0.16.1
dataclasses==0.8
decorator==4.0.11
defusedxml==0.5.0
dill==0.2.7.1
docutils==0.14
dronekit==2.9.2
entrypoints==0.2.3
et-xmlfile==1.0.1
eventlet==0.22.0
fastcache==1.0.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.16.0
gensim==3.4.0
glfw==1.8.0
greenlet==0.4.12
gym==0.9.6
h5py==2.7.1
heapdict==1.0.0
html2text==2018.1.9
html5lib==0.9999999
idna==2.6
imageio==2.1.2
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.0.5
itsdangerous==0.24
jdc==0.0.8
jdcal==1.3
jedi==0.10.2
jieba==0.39
Jinja2==3.0.3
jmespath==0.9.3
joblib==0.11
json5==0.8.5
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==1.0.9
jupyterlab-server==1.0.0
Keras==2.0.9
librosa==0.5.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==2.0.1
matplotlib==2.1.0
mistune==0.8.3
mock==2.0.0
monotonic==1.3
more-itertools==7.0.0
moviepy==0.2.3.2
msgpack==0.5.6
msgpack-python==0.4.8
mujoco-py==1.50.1.59
nbconvert==5.4.0
nbformat==4.4.0
nbzip==0.1.0
networkx==1.11
nltk==3.2.5
notebook==5.7.0
numba==0.35.0
numexpr==2.6.4
numpy==1.12.1
oauthlib==2.0.6
olefile==0.44
openpyxl==2.5.0b1
packaging==16.8
pandas==0.23.3
pandocfilters==1.4.1
parso==0.4.0
partd==0.3.8
patsy==0.4.1
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.2.0
pinyin==0.4.0
plotly==2.0.15
pluggy==0.11.0
pomegranate==0.9.0
prettytable==0.7.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.0
psycopg2==2.7.4
ptyprocess==0.5.2
py==1.8.0
py4j==0.10.7
pyasn1==0.4.4
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.18
pycurl==7.43.0
pydot==1.2.4
pydotplus==2.0.2
pyglet==1.3.0
Pygments==2.2.0
pygpu==0.7.5
PyJWT==1.5.3
pymavlink==2.2.20
pymc3==3.2
PyOpenGL==3.1.0
pyOpenSSL==17.4.0
pyparsing==2.2.0
pyrsistent==0.15.2
PySocks==1.6.8
pyspark==2.4.3
pytest==4.5.0
python-crfsuite==0.9.2
python-dateutil==2.6.1
python-docx==0.8.11
python-engineio==2.0.2
python-socketio==1.8.4
python-speech-features==0.6
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.1.2
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
resampy==0.2.0
rsa==3.4.2
ruamel-yaml==0.11.14
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.19.1
scipy==1.2.1
seaborn==0.8.1
SeekWell==0.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
SimpleWebSocketServer==0.1.1
six==1.11.0
smart-open==1.5.6
snownlp==0.12.3
SoundFile==0.9.0.post1
SQLAlchemy==1.1.13
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.0
tensorflow==1.3.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1
toolz==0.8.2
torch==0.4.0
torchvision==0.2.1
tornado==4.5.3
tqdm==4.11.2
traitlets==4.3.2
tweepy==3.5.0
twython==3.6.0
udacity-pa==0.2.9
urllib3==1.22
utm==0.4.2
uvloop==0.8.1
vega==0.4.4
vincent==0.4.4
wcwidth==0.1.7
webencodings==0.5
websockets==4.0.1
Werkzeug==2.0.3
widgetsnbextension==3.1.0
wordcloud==1.4.1
wptools==0.4.7
xlrd==1.1.0
